+++ lib/src/native/realm_core.dart
@@ -165,6 +165,11 @@ class _RealmCore {
         _realmLib.realm_config_set_data_initialization_function(configHandle._pointer, Pointer.fromFunction(initial_data_callback, FALSE), config.toGCHandle());
       }
 
+      if (config.shouldCompactCallback != null) {
+        _realmLib.realm_config_set_should_compact_on_launch_function(
+            configHandle._pointer, Pointer.fromFunction(should_compact_callback, 0), config.toGCHandle());
+      }
+
       return configHandle;
     });
   }
@@ -183,6 +188,16 @@ class _RealmCore {
     }
 
     return FALSE;
+  }
+
+  static int should_compact_callback(Pointer<Void> userdata, int totalSize, int usedSize) {
+    final Configuration? config = userdata.toObject();
+    if (config == null) {
+      return 0;
+    }
+    config.shouldCompactCallback!(totalSize, usedSize);
+
+    return 1;
   }
 
   SchedulerHandle createScheduler(int isolateId, int sendPort) {
@@ -652,7 +667,8 @@ class _RealmCore {
       final handle = AppConfigHandle._(_realmLib.realm_app_config_new(app_id, httpTransport._pointer));
 
       _realmLib.realm_app_config_set_base_url(handle._pointer, configuration.baseUrl.toString().toUtf8Ptr(arena));
-      _realmLib.realm_app_config_set_default_request_timeout(handle._pointer, configuration.defaultRequestTimeout!.inMilliseconds);
+
+      _realmLib.realm_app_config_set_default_request_timeout(handle._pointer, configuration.defaultRequestTimeout.inMilliseconds);
 
       if (configuration.localAppName != null) {
         _realmLib.realm_app_config_set_local_app_name(handle._pointer, configuration.localAppName!.toUtf8Ptr(arena));